tidy(ui): move staging area components to correct dir
