+++ invokeai/frontend/web/src/features/controlLayers/components/StagingArea/QueueItemProgressImage.tsx
