+++ invokeai/frontend/web/src/features/controlLayers/components/StagingArea/context.tsx
