+++ invokeai/frontend/web/src/features/controlLayers/components/StagingArea/shared.ts
