+++ invokeai/frontend/web/src/features/controlLayers/components/StagingArea/state.ts
